requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
torch>=1.10.0
tqdm>=4.62.0
networkx>=2.6.0
python-igraph>=0.9.0
scikit-learn>=1.0.0
causal-learn>=0.1.3
cdt>=0.6.0
pgmpy>=0.1.20
salesforce-causalai>=1.0.0
typing_extensions>=4.0.0

# NOTEARS algorithm (install from GitHub if not available on PyPI)
# pip install git+https://github.com/xunzheng/notears.git
# OR if using a local copy, ensure notears package is in PYTHONPATH
